Reimplement whitespace checker via an external job
The search() function took too long to check very large files (>10k
lines). The problem is that usage of a timer doesn't really make the
entire thing asynchronous, so I had to use an external tool through the
job API. The timer is still used and the processing functions still
block, but it isn't noticeable even with files with over 100k lines. For
the interval, one second seems reasonable, but I haven't had a problem
even with just a tenth.

The result is now stored in a buffer-local variable, which means that
different windows get the correct value in their statuslines. The
statusline fragment is now completely disabled when in insert mode,
because the previous behavior didn't really solve the problem, if
another trailing whitespace was on another line after the one currently
edited (the fragment would appear and disappear as one appended spaces
to the current line).

The statusline function has been renamed, because it doesn't show only
ALE stuff anymore.