Improve lockscreen
The lockscreen now shows up immediately (although with an old image or a
black screen). The screen goes black after seven seconds once , to catch
people who do something with my PC. This could be improved by turning
off the screen in lock_generator too and logging, if the screen is on or
off.