Fix cursor when entering CocList
The cursor worked kind of weird. It had a highlight group set, but it
didn't actually do anything until I used CocList (or similar). Now, the
cursor has no special highlighting so it always stays the color of the
current character it's on (or rather the inverse). This functionality is
probably implemented by the terminal rather than vim.